build: rename payload generator to kocolor-compiler
This commit renames the Rust-based payload generation tool to `kocolor-compiler` and consolidates the production build pipeline. The change aligns the binary naming with the project's documentation and transitions the compiler from a simple CLI stub to a full-featured enrichment and signing engine.

- **Binary Renaming and Integration**:
    - Renamed `src/bin/generate_payload.rs` to `src/bin/kocolor-compiler.rs`.
    - Updated `Cargo.toml` to reflect the new binary name and source path.
    - Updated `runMe.sh` to call the renamed binary during the build process.
- **Compiler Logic Consolidation**:
    - Replaced the previous basic CLI logic in `kocolor-compiler.rs` with the comprehensive batch processing pipeline, including Zstd compression, Ed25519 signing, and SHA-256 hashing.
    - Maintained support for compile-time enrichment, specifically CIELAB colorimetry conversion and chemistry phase mapping for cosmetic items.
    - Preserved the generation of both the `search_index.json` and the signed `manifest.json`.
- **Documentation Alignment**:
    - Updated `ComputeColorImplementation.md` and `Implementation_V1_Core.md` to reference `kocolor-compiler` as the authoritative tool for package normalization and enrichment.
